refactor: move entities folder to core folder
(cherry picked from commit 249a4ad1288515e6f81fde5a40f0c42ef3efbf09)
