+++ backend/src/core/entities/user.entity.ts
