+++ backend/src/users/users.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { UsersController } from './users.controller';
 import { UsersService } from './users.service';
-import { User } from '../entities/user.entity';
+import { User } from '../core/entities/user.entity';
 
 @Module({
   imports: [TypeOrmModule.forFeature([User])],
